refactored media finder code out of generator, refactored generator config, removed config factory...
