+++ src/MediaFinder.ts
@@ -1,0 +1,4 @@
+
+export default interface MediaFinder {
+    copyMedia() : Promise<any>;
+}
